Refactor: calcPaymentToBeUsed 함수 util에서 삭제
- WalletProvider로 이동
