+++ src/components/orderArea/PutBtn.jsx
@@ -2,21 +2,19 @@ import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import { Button } from 'components/orderArea/PutBtn.style';
 import { TIME_TO_SELCT_PRODUCT } from 'constant/constant';
-import { calcPaymentToBeUsed } from 'utils/util';
 import useVMState from 'hooks/useVMState';
 import { FinalPayContext, FinalPaySetContext } from 'contexts/FinalPayProvider';
 import { HistoryDispatchContext } from 'contexts/HistoryProvider';
-import { WalletContext, WalletSetContext } from 'contexts/WalletProvider';
+import { WalletSetContext } from 'contexts/WalletProvider';
 
 export default function PutBtn({ inputPay }) {
   const [finalPay, setFinalPay] = [useContext(FinalPayContext), useContext(FinalPaySetContext)];
   const { addInputHistory } = useContext(HistoryDispatchContext);
-  const walletState = useContext(WalletContext);
+  const { calcPaymentToBeUsed, decreaseUnitsToBeUsed } = useContext(WalletSetContext);
   const { startTimerToReset } = useVMState();
-  const { decreaseUnitsToBeUsed } = useContext(WalletSetContext);
 
   const handlePutBtnClick = () => {
-    const paymentToBeUsed = calcPaymentToBeUsed(walletState, inputPay);
+    const paymentToBeUsed = calcPaymentToBeUsed(inputPay);
     const totalPay = finalPay + paymentToBeUsed.total;
     setFinalPay(totalPay);
     decreaseUnitsToBeUsed(paymentToBeUsed.unitsToBeUsed);
